Add some checking and implements role model

+++ src/main/java/ru/rtstudy/educplatformsecurity/repository/DifficultRepository.java
@@ -5,10 +5,12 @@ import org.springframework.data.jpa.repository.Query;
 import ru.rtstudy.educplatformsecurity.model.Difficult;
 import ru.rtstudy.educplatformsecurity.model.constant.DifficultLevel;
 
-public interface DifficultRepository extends JpaRepository<Difficult, Byte> {
+public interface DifficultRepository extends JpaRepository<Difficult, Long> {
 
     @Query("""
-            select d from Difficult d where d.difficultLevel = :difficult
+            select d 
+            from Difficult d 
+            where d.difficultLevel = :difficult
             """)
     Difficult getDifficultByDifficultName(DifficultLevel difficult);
 }
